+++ CycleRoute/client/src/components/Tabs/Bike Stations/bikeStations.css
@@ -1,0 +1,8 @@
+/* Aligns all post "cards" in a grid,limits of post going horizontally */
+.allstations {
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  grid-column-gap: 20px;
+  grid-row-gap: 20px;
+  margin: 1rem;
+}
